Split index.tsx into AppDispatcher and GameScreen
index.tsx is now a thin dispatcher that reads first-launch state from
storage and replaces to /game, passing a firstTime param on first launch.
game.tsx holds all former game logic and uses the param to decide whether
to push /help (first launch) or /menu (returning user).
