+++ src/app/_layout.tsx
@@ -36,6 +36,7 @@ export default function RootLayout() {
               }}
             />
             <Stack.Screen name="index" />
+            <Stack.Screen name="game" />
             <Stack.Screen
               name="help"
               options={{
